+++ app/src/main/res/layout/activity_joystick.xml
@@ -10,12 +10,13 @@
 
     <io.github.controlwear.virtual.joystick.android.JoystickView xmlns:custom="http://schemas.android.com/apk/res-auto"
         android:id="@+id/joystick_view"
-        android:layout_width="180dp"
-        android:layout_height="180dp"
+        android:layout_width="196dp"
+        android:layout_height="196dp"
         android:background="@drawable/robosoccer_joy"
         custom:JV_buttonColor="#da0004"
         custom:JV_buttonSizeRatio="20%"
         custom:JV_fixedCenter="false"
+        android:layout_margin="8dp"
         custom:layout_constraintBottom_toBottomOf="parent"
         custom:layout_constraintEnd_toEndOf="parent" />
 
@@ -36,8 +37,8 @@
 
     <ImageView
         android:id="@+id/controls_swap_button"
-        android:layout_width="48dp"
-        android:layout_height="48dp"
+        android:layout_width="64dp"
+        android:layout_height="64dp"
         android:layout_marginTop="16dp"
         android:background="@drawable/circle"
         android:padding="4dp"
@@ -62,6 +63,7 @@
         app:layout_constraintTop_toTopOf="parent" />
 
     <ImageView
+        android:elevation="16dp"
         android:id="@+id/button_view"
         android:layout_width="88dp"
         android:layout_height="88dp"
