+++ app/src/main/res/layout/layout_toolbar.xml
@@ -12,8 +12,8 @@
         android:layout_height="wrap_content"
         android:layout_gravity="center"
         android:fontFamily="sans-serif-condensed"
-        android:text="RoboEvents"
-        android:textAppearance="@style/Base.TextAppearance.AppCompat.Large"
+        android:text="@string/app_name"
+        android:textSize="18sp"
         android:textColor="#000"
         android:textStyle="bold" />
